Change IntegralImage to use float instead of double; initialise matrix in constructor instead of in compute().

+++ code/src/processors/integral_image.h
@@ -17,7 +17,7 @@ class IntegralImage
 public:
   IntegralImage(Mat img);
   void compute();
-  double area(Point start, Point end);
+  float area(Point start, Point end);
 private:
   Mat m_img;
   Mat m_int;
